+++ lib/hashie/logger.rb
@@ -1,9 +1,16 @@
+require 'logger'
+
 module Hashie
   # The logger that Hashie uses for reporting errors.
   #
   # @return [Logger]
   def self.logger
-    @logger ||= Logger.new(STDOUT)
+    @logger ||=
+      if defined?(::Rails)
+        Rails.logger
+      else
+        Logger.new(STDOUT)
+      end
   end
 
   # Sets the logger that Hashie uses for reporting errors.